Remove aux grid stuff
