+++ amrex_code/tests/cyclotron_borders.input
@@ -1,17 +1,18 @@
-# Example of electron slowing down in E-field
-# Should come to a complete stop
+# Example of electron cyclotron motion
+# Should complete one revolution
 # SI-inputs
-# E = 1MV/m
+# B = 0.55 T
 # m = m_e
 # q = q_e
 # dx = 2.773e-5
 # vel = 0.01 c
 
+
 n_cell =  12 12 12
-max_grid_size = 3 3 3
+max_grid_size = 6 6 3
 x_periodic = 1
 
-nsteps = 3680
+nsteps = 6000
 start_step = 0 # If this is not zero we will try to do a restart from a plotfile
 output_interval = -1 ;
 checkpoint_interval = -1 ;
@@ -20,8 +21,8 @@ data_folder_name = "Single_Particle_Data_out"
 
 q = -1.60217662e-19
 m = 9.427127615688092e-16
-pos = 6 6 6
-vel = 0.1 0.0 0.1
+pos = 12 11 6
+vel = 0.01 0.0 0.01
 
-E_init = 0.31929995744680856 0 0
-B_init = 0 0 0
+E_init = 0 0 0
+B_init = 0 0 58.8395
