+++ amrex_code/tests/energy_conservation.cpp
@@ -100,12 +100,10 @@ void main_main()
     amrex::IntVect dom_hi(AMREX_D_DECL(n_cell[X]-1, n_cell[Y]-1, n_cell[Z]-1));
     amrex::Box domain(dom_lo, dom_hi,typ);
     amrex::BoxArray ba(domain);
-    amrex::BoxArray gba(domain);
 
     // Initialize the boxarray "ba" from the single box "bx"
     // Break up boxarray "ba" into chunks no larger than "max_grid_size" along a direction
     ba.maxSize({max_grid_size[X],max_grid_size[Y],max_grid_size[Z]});
-    gba.maxSize({max_grid_size[X],max_grid_size[Y],max_grid_size[Z]});
     // This defines the physical box, [-1,1] in each direction.
     amrex::RealBox real_box({AMREX_D_DECL(0,0,0)},
                      {AMREX_D_DECL((double)n_cell[X] , (double)n_cell[Y],(double)n_cell[Z])});
@@ -116,17 +114,11 @@ void main_main()
     amrex::DistributionMapping dm(ba);
     CParticleContainer P(geom,dm,ba);
     //distribute_processes_pdens(dm,geom,ba,bernstein_density,"SFC");    
-    shift_and_grow<X>(gba,Nghost);
-    shift_and_grow<Y>(gba,Nghost);
-    shift_and_grow<Z>(gba,Nghost);
-    auto gdomain=gba.minimalBox(); 
-    amrex::Geometry ggeom(gdomain,&real_box,amrex::CoordSys::cartesian,is_periodic.data());
 
 
-    amrex::MultiFab E_L(gba,dm,Ncomp,Nghost); 
     amrex::MultiFab E(ba,dm,Ncomp,Nghost);
     amrex::MultiFab B(ba,dm,Ncomp,Nghost);
-    auto SimIO=SimulationIO(geom,ggeom,gba,E,B,P,dt,data_folder_name);
+    auto SimIO=SimulationIO(geom,E,B,P,dt,data_folder_name);
 
 
 ////////////7
@@ -162,7 +154,7 @@ for(int step=start_step; step<nsteps;step++){
     if(step % checkpoint_interval ==0 && checkpoint_interval  !=-1){
         SimIO.write(step,true,false);
     }
-    Theta_map<1>(geom,ggeom,P,E,E_L,B,dt);
+    Theta_map<1>(geom,P,E,B,dt);
 }
